Migrate sessions to JWT

+++ server/src/controllers/carts.controller.js
@@ -1,4 +1,4 @@
-import { cartService } from "../services/index.js";
+import { userService, cartService } from "../services/index.js";
 import response from "../helpers/response.js";
 
 export const getCarts = async (req, res, next) => {
@@ -24,7 +24,7 @@ export const getCartByID = async (req, res, next) => {
 
 export const getCart = async (req, res, next) => {
     try {
-        const { cart } = req.session.user || { cart: null };
+        const { cart } = await userService.getUserById(req.uid) || { cart: null };
         const result = await cartService.getCart(cart);
         res.send(await response('success', result, req.query))
     } catch (error) {
@@ -47,7 +47,8 @@ export const addCart = async (req, res, next) => {
 export const addProductToCart = async (req, res, next) => {
     try {
         const { pid } = req.params
-        const { cart: cid } = req.session.user
+        const { cart: cid } = await userService.getUserById(req.uid) || { cart: null };
+        if (!cid) throw new Error('Cart not found')
         const result = await cartService.addProductToCart(pid, cid)
         res.send(await response('success', result, req.query))
     } catch (error) {
@@ -59,7 +60,8 @@ export const addProductToCart = async (req, res, next) => {
 export const deleteProductFromCart = async (req, res, next) => {
     try {
         const { pid } = req.params
-        const { cart: cid } = req.session.user
+        const { cart: cid } = await userService.getUserById(req.uid) || { cart: null };
+        if (!cid) throw new Error('Cart not found')
         const result = await cartService.deleteProductFromCart(pid, cid)
         res.send(await response('success', result, req.query))
     } catch (error) {
@@ -84,7 +86,8 @@ export const updateProductUnits = async (req, res, next) => {
     try {
         const { pid } = req.params
         const { units } = req.body
-        const { cart: cid } = req.session.user
+        const { cart: cid } = await userService.getUserById(req.uid) || { cart: null };
+        if (!cid) throw new Error('Cart not found')
         const result = await cartService.updateProductUnits(pid, cid, units)
         res.send(await response('success', result, req.query))
     } catch (error) {
